[버그]  Pathvariable id명시 (#69)
* fix : controller의 pathvariable id값 명시

* fix : postService likeCount return 변경

+++ service-post/src/main/java/com/team13/servicepost/util/RandomPostGenerator.java
@@ -261,7 +261,7 @@ public class RandomPostGenerator {
     }
 
     public static Long likesCount() {
-        return random.nextLong(50);
+        return (long) random.nextInt(50);
     }
